feat: Standarize the string formating of sum types and values (#2432)
The Sum value and type helpers get coerced back into `val.Sum` and
`tys.Sum` after a serialization roundtrip. This used to change the
rendering of the values, falling back to the verbose `Sum(tag=#,
typ=[..., ...], vals=[...,...])`.

This PR centralizes the str/repr definition, so equivalent values and
types are always rendered in the same way. (I needed this for roundtrip
checks).