fix: Fixed export of `Call` and `LoadConst` nodes in `hugr-py`. (#2429)
`Call` and `LoadConst` nodes have static inputs which were erroneously
exported in `hugr-py`. The function to be called and the constant to be
loaded are arguments to the operations in `hugr-model`, so the static
input isn't needed. Before this fix, this caused import errors when the
same function would be called twice.