feat: Names of private functions become `core.title` metadata. (#2448)
This PR changes the way that function names are exported/imported
between `hugr-core`/`hugr-py` on the one side and `hugr-model` on the
other. When a function is public, the exported `hugr-model` symbol name
will match the `hugr-core` name. For private functions, the `hugr-model`
exported symbol name will be `_{node_id}` and the `hugr-core` name is
preserved via `core.title` metadata. On import, the `hugr-core` name is
the title metadata if it is present; otherwise it is the `hugr-model`
symbol name.