feat: Detect and fail on unrecognised envelope flags (#2453)
When reading an envelope header flags, we ignored any unrecognised bits.

This change adds a new error to both the python and rust decoders, that
ensures we detect and report this unexpected flags (nudging the user to
update their HUGR version).

This will let us add new config flags in the future that change the
payload format, having older versions gracefully fail to read them.