feat(py): Use SumValue serialization for tuples (#2466)
The python lib had a `_serialization.ops.TupleValue` struct for special
cases of single-variant sums, that got encoded with an special tag.
The rust does not have the specialized varinat, and simply loads it as
`SumValue`.

This PR removes the unnecessary special case from the python side, to
simplify the serialization logic and help with the roundtrip checks.

The json schemas have been updated, but the change is non-breaking.