ci: Update release-plz changelog header (#2474)
The default seems to have changed to 
```toml
header = """# Changelog

All notable changes to this project will be documented in this file.

The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).

## [Unreleased]

"""
```

Here we change it back to just the `# Changelog` line