feat: Support LineConvexChecker (#2487)
This PR's goal is to expose the new `LineConvexChecker` in `portgraph`
in the hugr API. For details see the original PR:
https://github.com/CQCL/portgraph/pull/240

This PR will require a (non-breaking) portgraph release. Currently
pinning portgraph main (to be reverted).

There are also two functions whose API could be generalised -- but that
would be a breaking change. I've marked them as TODOs in the code.

## Breaking change?

Confirm: hiking the patch version of portgraph is considered
non-breaking, right?

The current "breaking change" detected by semver-checks is due to
`struct TopoConvexChecker` now being a type synonym for
`ConvexChecker<'g, Base, pg:TopoConvexChecker<CheckerRegion<'g,
Base>>>`. It behaves exactly identically from a user's perspective, so I
would be in favour of ignoring that warning, unless someone can spot an
issue with it?

Alternatively, I could define a newtype `struct
TopoConvexChecker(ConvexChecker)` that delegates all methods to the
inner struct. Presumably semver-checks would accept that as non-breaking
(but it would be ugly, verbose and would need to be undone before the
next breaking release).