feat: SiblingSubgraph supports function calls (#2528)
ended up being a bit more work than expected! With this PR,
`SiblingSubgraph` supports incoming boundary static edges that link a
function declared/defined outside of the subgraph to a call within the
subgraph.

I've taken the opportunity to clean up the docs and error handling for
any non-value (and non-function static edges).