feat: Loosen bound on Patch trait (#2545)
This generalisation is useful in tket2, where structs such as `Circuit`
and `ResourceScope` do not implement `HugrView` but for which
implementations of `Patch` make sense.

Ideally, one would also loosen `PatchVerification` to be similarly
parametrised on `H: NodesIter`, but that would be a breaking change (and
@acl-cqc might have an opinion on that latter part). I think we should
go ahead with this for the time being, have created issue
https://github.com/CQCL/hugr/issues/2546 for the changes to
`PatchVerification`.