feat: add_input/output for arbitrary DFGBuilders (#2564)
These methods were only available on `FuncDefn`s.

This PR moves the definition to `DFGBuilder`, so we can add
input/outputs to DFGs and DataflowBlocks too.
Other unsupported ops return an Error, we can extend the support in the
future.

The methods are re-exported from `DfgWrapper`, but I had to `.unwrap`
the errors for now to avoid API breakage. We should remove those in the
next breaking release.