test: Add framework for LLVM execution tests involving panics (#2568)
The default LLVM lowering for panic ops just aborts the process. This
makes testing of panic behaviour (e.g. for the new borrow array ops)
impossible. This PR adds an alternative `TestPanicPreludeCodegen` that
calls into a small panic runtime to safely exit programs and report the
error message, instead of aborting the process.