feat(persistent): Redesign CommitStateSpace, bound Commit lifetime (#2534)
This PR updates the relrc dependency to v0.5.0 and completely redesigns
the `CommitStateSpace` struct. The result is much cleaner:

- `PersistentHugr` is an immutable data type following the design of the
types of the [`im` crate](https://docs.rs/im/latest/im/), i.e. it is a
mutable Rust type that is cheap to clone and mutate (immutability -- if
desired -- is achieved by cloning before mutating).
- `Commit` is a "transaction", i.e. one atomic mutation that was applied
to a `PersistentHugr`. It has an ID, obtainable through `commit.id()`
and a lifetime parameter that ensures it does not stay in scope beyond
the lifetime of the `PersistentHugr`
- Each `PersistentHugr` belongs to a unique `CommitStateSpace`. A
`CommitStateSpace` is a registry of all commits of the `PersistentHugr`s
within that state space. It keeps the map between commits and IDs.

This is a breaking change, but does not trigger a breaking HUGR release
because it is hidden behind an unstable feature.