fix: validation outside entrypoint, normalize_cfgs w/ nonlocal edges (#2633)
fixes #2598 

The bug in validation was simple, `validate_subtree` was only called
beneath the entrypoint. Change that...

Of course there were test failures - Hugrs we thought were valid but
actually weren't.

* Many of these were `ReplaceTypes` tests on DFG-entrypoint Hugrs:
`ReplaceTypes` acted only on/beneath the entrypoint, so (the
Input/Output of) the FuncDefn containing the DFG no longer matches it.
I've changed these to make the FuncDefn be entrypoint, without a DFG, so
all relevant nodes are ReplaceType'd.

* NormalizeCFGs can move the Entry block of a CFG into a DFG outside it.
If there were nonlocal `Dom` edges from inside that entry block, this
would make the Hugr invalid. (Thus, a second bug, which we had not
noticed because of the validation bug.) I've fixed this by only
transforming if there are no `Dom` edges from the entry block.

* A similar case (when the predecessor of the Exit block can be moved
outside the CFG) is now also done only if there are no *incoming* `Dom`
edges.