feat: add hugr-core StaticGraph, deprecate hugr-passes CallGraph (#2698)
Respin of #2531, marking the node/edge enums in the new StaticGraph as
`non-exhaustive` and adding Consts + LoadConstants.

(Lukas originally planned to make something like this to reflect
references from Terms to functions, and non-exhaustive allows extending
this StaticGraph for that purpose - if we do go that way, however, I
have a different plan, can describe if relevant. Rather I think
non-exhaustive is just a reasonable allowance for future extension to
new use-cases hopefully without damaging existing use-cases.)

This is taken from #2555, i.e. which uses it (although there it is still
called CallGraph in hugr-core).

Naming: could be ModuleGraph ?? Other ideas welcome.

I will add a test with some Consts + use of `out_edges` (tho will be
supplied in #2555 !)