feat: ReplaceTypes: recurse on replacements, much deprecation (#2442)
closes #2668 (by a combination of the two routes, minus the extension).

There are two(/3) goals here:
* firstly, to improve compositionality - if you configure ReplaceTypes
to convert type A to type B, and also type B to type C, then if it sees
an A it "should" produce a C. (This is not the case at present, you'll
get a B.) Likewise for op-replacements containing ops that themselves
would be replaced.
* secondly, to allow lowering "drop" without creating an invalid hugr -
instead, you can now call `copy_discard_op(ty, 0)`, as demonstrated in
the test.
* finally, as a driveby, fixes bug where linearization was applied
(only) if the op *or any previous op in the subtree* had been changed

Note that the first is a breaking behaviour change (potentially very
breaking, i.e. infinite loop), so to keep this PR non-breaking, I add
new methods `set_replace(_parametrized,)_(op,type)` that process the RHS
recursively. To satisfy the second goal, the op-replacing callback is
passed the ReplaceTypes (hence rolling two breaking signature changes
into one PR).. The naming of the new methods is a little odd, largely
because the first choice `replace_blah` is already taken, and we keep
those methods (with non-recursive behaviour) but deprecated.

(If later we really want to keep non-recursive replacement, we can add
that later, but there are enough methods right now!)

Note for consts, there's no real equivalent to "recursive" replacement,
because there is no way for the ReplaceTypes to look inside CustomConsts
like lists, they are opaque. Instead, recursion must be done manually on
the elements, as per e.g. `array_const`:
https://github.com/CQCL/hugr/blob/fcd102645c0af644a29d55453e97ee9894d9aa69/hugr-passes/src/replace_types/handlers.rs#L75