feat: Make `ComposablePass.__call__` return a Hugr (#2697)
Implementing the revisions to `ComposablePass` and `ComposedPass` as
discussed with Agustin.

I still find the `_apply` and `_apply_inplace` rather unintuitive
(although now I realise that it will work). Looking at the default
implementations in the `ComposablePass` protocol these look circular but
as you're always overrididing at least one of the `_apply` or
`_apply_inplace` methods whenever you implement the protocol.


Would be interested to know if there are suggestions to improve this.

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>
Co-authored-by: Agustín Borgna <agustin.borgna@quantinuum.com>