refactor: Direct import of model representation to Python (#2683)
Closes #2287 .

Testing:

* All existing hugr-py tests that call `validate()` perform a
round-trip, checking that the hashes of the start and end Hugrs
(computed using the `_NodeHash` method defined in `conftest.py`) agree.
* I exported all the Hugrs generated by tests in the guppylang repo
(using `just export-integration-tests`) and tried importing them (1)
using the old method via json and (2) using the new method, and checked
that the hashes of the two imported Hugrs agree (using the same
`_NodeHash` method).

There are a few TODO comments remaining in `load.py` which I could not
see a way to resolve using the existing model. I will investigate these
further and raise issues if necessary.

---------

Co-authored-by: Lukas Heidemann <lukas@heidemann.me>